requirements.txt
Flask
mysql-connector-python
bcrypt
reportlab
flask-cors